# api/routes.py
import os
import glob
import time
from uuid import UUID
from flask import Blueprint, request, jsonify, send_file
from marshmallow import ValidationError
from app import db
from models import Download
from api.schemas import VideoInfoSchema, DownloadSchema
from utils.downloader import get_cached_video_info, get_cached_formats, start_download_task
import logging

logger = logging.getLogger(__name__)

api_bp = Blueprint('api', __name__)

@api_bp.route('/info', methods=['GET'])
def get_info():
    """Get basic video metadata without formats"""
    url = request.args.get('url')
    if not url:
        return jsonify({'error': 'URL parameter is required'}), 400

    try:
        info = get_cached_video_info(url)
        return jsonify(info)
    except Exception as e:
        logger.error(f"Error getting video info: {str(e)}")
        return jsonify({'error': str(e)}), 400

@api_bp.route('/formats', methods=['GET'])
def get_formats():
    """Get available video formats"""
    url = request.args.get('url')
    if not url:
        return jsonify({'error': 'URL parameter is required'}), 400

    try:
        filtered = request.args.get('filtered', 'false').lower() == 'true'
        logger.debug(f"Getting formats with filtered={filtered}")
        formats = get_cached_formats(url, filtered=filtered)
        return jsonify(formats)
    except Exception as e:
        logger.error(f"Error getting video formats: {str(e)}")
        return jsonify({'error': str(e)}), 400

@api_bp.route('/download', methods=['GET'])
def create_download():
    """Create download task"""
    try:
        url = request.args.get('url')
        if not url:
            return jsonify({'error': 'URL parameter is required'}), 400

        format_id = request.args.get('format')
        video_format_id = request.args.get('video_format_id')
        audio_format_id = request.args.get('audio_format_id')
        
        if not format_id and (not video_format_id or not audio_format_id):
            return jsonify({'error': 'Either format or both video_format_id and audio_format_id are required'}), 400
            
        # Get video info for format validation
        video_info = get_cached_formats(url)
        
        if format_id:
            # Check single format existence
            selected_format = None
            for f in video_info:
                if f.get('format_id') == format_id:
                    selected_format = f
                    break
            
            if not selected_format:
                return jsonify({'error': f'Invalid format ID: {format_id}'}), 400
            
            task_id = UUID(bytes=os.urandom(16))
            download = Download(
                task_id=task_id,
                url=url,
                format=format_id
            )
        else:
            # Check video and audio format existence
            video_format = None
            audio_format = None
            for f in video_info:
                if f.get('format_id') == video_format_id and f.get('vcodec') != 'none':
                    video_format = f
                elif f.get('format_id') == audio_format_id and f.get('acodec') != 'none':
                    audio_format = f
            
            if not video_format:
                return jsonify({'error': f'Invalid video format ID: {video_format_id}'}), 400
            if not audio_format:
                return jsonify({'error': f'Invalid audio format ID: {audio_format_id}'}), 400
            
            task_id = UUID(bytes=os.urandom(16))
            download = Download(
                task_id=task_id,
                url=url,
                video_format=video_format_id,
                audio_format=audio_format_id
            )
            
        db.session.add(download)
        db.session.commit()
        
        # Start async download
        if format_id:
            start_download_task(str(task_id), url, format_id=format_id)
        else:
            start_download_task(str(task_id), url, video_format_id=video_format_id, audio_format_id=audio_format_id)
        
        # Prepare response
        response = {
            'task_id': str(download.task_id),
            'url': download.url,
            'created_at': download.created_at.isoformat()
        }

        # Add format specific information
        if format_id:
            response.update({
                'format': download.format,
                'format_info': {
                    'format': selected_format.get('format'),
                    'ext': selected_format.get('ext'),
                    'resolution': selected_format.get('resolution'),
                    'filesize': selected_format.get('filesize'),
                    'filesize_mb': round(selected_format.get('filesize', 0) / 1024 / 1024, 2) if selected_format.get('filesize') else None,
                    'filesize_approx': selected_format.get('filesize_approx'),
                    'filesize_approx_mb': round(selected_format.get('filesize_approx', 0) / 1024 / 1024, 2) if selected_format.get('filesize_approx') else None
                }
            })
        else:
            response.update({
                'video_format': download.video_format,
                'audio_format': download.audio_format,
                'format_info': {
                    'video': {
                        'format': video_format.get('format'),
                        'ext': video_format.get('ext'),
                        'resolution': video_format.get('resolution'),
                        'filesize': video_format.get('filesize'),
                        'filesize_mb': round(video_format.get('filesize', 0) / 1024 / 1024, 2) if video_format.get('filesize') else None,
                        'filesize_approx': video_format.get('filesize_approx'),
                        'filesize_approx_mb': round(video_format.get('filesize_approx', 0) / 1024 / 1024, 2) if video_format.get('filesize_approx') else None
                    },
                    'audio': {
                        'format': audio_format.get('format'),
                        'ext': audio_format.get('ext'),
                        'filesize': audio_format.get('filesize'),
                        'filesize_mb': round(audio_format.get('filesize', 0) / 1024 / 1024, 2) if audio_format.get('filesize') else None,
                        'filesize_approx': audio_format.get('filesize_approx'),
                        'filesize_approx_mb': round(audio_format.get('filesize_approx', 0) / 1024 / 1024, 2) if audio_format.get('filesize_approx') else None
                    }
                }
            })
            
        return jsonify(response), 202

    except ValidationError as err:
        return jsonify({'error': err.messages}), 400
    except Exception as e:
        logger.error(f"Error creating download: {str(e)}")
        return jsonify({'error': str(e)}), 500

@api_bp.route('/download/<task_id>', methods=['GET'])
def get_download_status(task_id):
    """Get download task status"""
    try:
        task_uuid = UUID(task_id)
        download = Download.query.filter_by(task_id=task_uuid).first()
        if not download:
            return jsonify({'error': 'Download task not found'}), 404

        result = DownloadSchema().dump(download)
        if result.get('file_path'):
            # Add full HTTPS download URLs if file exists
            host = request.host
            result['download_url'] = f"https://{host}/api/download/{task_id}/file"
            
            # Add direct file URL with extension if download is completed
            if download.status == 'completed':
                # Get file extension based on format
                ext = 'mp4'  # default extension
                if download.format:
                    # Get format info from video info
                    video_info = get_cached_formats(download.url)
                    for f in video_info:
                        if f.get('format_id') == download.format:
                            ext = f.get('ext', 'mp4')
                            break
                result['file_url'] = f"https://{host}/api/download/{task_id}.{ext}"
            
            # Remove file_path from response since it's internal
            result.pop('file_path', None)
        return jsonify(result)
    except ValueError as e:
        logger.error(f"Invalid UUID format: {task_id}")
        return jsonify({'error': 'Invalid task ID format'}), 400

@api_bp.route('/download/<task_id>/file', methods=['GET'])
@api_bp.route('/download/<task_id>.<ext>', methods=['GET'])
def download_file(task_id, ext=None):
    """Download the completed file
    Supports two URL formats:
    - /api/download/{task_id}/file
    - /api/download/{task_id}.{ext}
    """
    logger.info(f"Download request received - task_id: {task_id}, ext: {ext}")
    
    try:
        # Convert task_id to UUID
        task_uuid = UUID(task_id)
        logger.debug(f"Converted task_id to UUID: {task_uuid}")
        
        # Get download record
        download = Download.query.filter_by(task_id=task_uuid).first()
        if not download:
            logger.error(f"Download task not found: {task_uuid}")
            all_tasks = Download.query.all()
            logger.debug(f"Available tasks: {[str(t.task_id) for t in all_tasks]}")
            return jsonify({'error': 'Download task not found'}), 404
            
        logger.info(f"Found download task: {download.task_id}, status={download.status}")
        
        # Verify download status
        if download.status == 'error':
            logger.error(f"Download failed: {download.error}")
            return jsonify({'error': f'Download failed: {download.error}'}), 400
        elif download.status != 'completed':
            logger.warning(f"Download not ready: status={download.status}, progress={download.progress}")
            return jsonify({
                'error': 'Download not completed yet',
                'status': download.status,
                'progress': download.progress
            }), 400
        
        # Setup paths
        downloads_dir = os.path.abspath('downloads')
        task_dir = os.path.join(downloads_dir, str(task_uuid))
        logger.debug(f"Looking for files in: {task_dir}")
        
        if not os.path.exists(task_dir):
            logger.error(f"Task directory not found: {task_dir}")
            return jsonify({'error': 'Download directory not found'}), 404
        
        try:
            # Check database file path first
            if download.file_path and os.path.exists(download.file_path):
                actual_file = download.file_path
                logger.info(f"Using file from database: {actual_file}")
            else:
                # Search in task directory
                logger.info(f"Searching for video files in: {task_dir}")
                video_files = []
                for ext in ['.mp4', '.mkv', '.webm', '.m4a']:
                    pattern = os.path.join(task_dir, f'*{ext}')
                    found = glob.glob(pattern)
                    if found:
                        logger.debug(f"Found files matching {pattern}: {found}")
                        video_files.extend(found)
                
                if not video_files:
                    logger.error(f"No video files found in {task_dir}")
                    return jsonify({'error': 'Video file not found'}), 404
                
                # Get most recent file
                video_files.sort(key=lambda x: os.path.getmtime(x), reverse=True)
                actual_file = video_files[0]
                logger.info(f"Selected most recent file: {actual_file}")
                
                # Update database
                download.file_path = actual_file
                db.session.add(download)
                db.session.commit()
            
            # Verify file
            file_stat = os.stat(actual_file)
            if file_stat.st_size == 0:
                logger.error(f"File is empty: {actual_file}")
                return jsonify({'error': 'File is empty'}), 500
            
            # Check for temporary files
            temp_files = []
            for pattern in ['*.part', '*.ytdl', '*.temp']:
                temp_files.extend(glob.glob(os.path.join(task_dir, pattern)))
            
            if temp_files:
                logger.warning(f"Found temporary files: {temp_files}")
                return jsonify({
                    'error': 'Download still in progress',
                    'status': 'downloading'
                }), 400
            
            # Check file stability
            initial_size = file_stat.st_size
            time.sleep(1)
            current_size = os.path.getsize(actual_file)
            
            if current_size != initial_size:
                logger.warning(f"File size changing: {initial_size} -> {current_size}")
                return jsonify({
                    'error': 'File still being written',
                    'status': 'processing'
                }), 400
            
            # Ensure file is readable
            if not os.access(actual_file, os.R_OK):
                logger.warning(f"Fixing permissions for {actual_file}")
                try:
                    os.chmod(actual_file, 0o644)
                except Exception as e:
                    logger.error(f"Failed to set file permissions: {e}")
                    return jsonify({'error': 'File not accessible'}), 403
            
            # Serve the file
            logger.info(f"Serving file: {actual_file} ({file_stat.st_size} bytes)")
            return send_file(
                actual_file,
                as_attachment=True,
                download_name=os.path.basename(actual_file)
            )
            
        except OSError as e:
            logger.error(f"OS error accessing file: {str(e)}", exc_info=True)
            return jsonify({'error': 'Error accessing file'}), 500
            
        except Exception as e:
            logger.error(f"Unexpected error: {str(e)}", exc_info=True)
            return jsonify({'error': 'Internal server error'}), 500
            
    except ValueError as e:
        logger.error(f"Invalid UUID format: {task_id}")
        return jsonify({'error': 'Invalid task ID format'}), 400